% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHCNdata-package.R
\docType{package}
\name{GHCNdata}
\alias{GHCNdata}
\title{Some Global Historical Climatological Network Data}
\description{
An R package with some data from the Global Historical Climatological Network.
    Also includes several functions for dealing with that data.
}
